LPD-40184 Rename to pasteItems
