+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/actions/pasteItems.ts
@@ -8,7 +8,7 @@ import {PASTE_ITEM} from './types';
 import type {LayoutData} from '../../types/layout_data/LayoutData';
 import type {FragmentEntryLink} from './addFragmentEntryLinks';
 
-export default function pasteItem({
+export default function pasteItems({
 	addedFragmentEntryLinks = [],
 	itemIds,
 	layoutData,
