+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/services/FragmentService.ts
@@ -259,7 +259,7 @@ export default {
 		);
 	},
 
-	pasteItem({
+	pasteItems({
 		itemIds,
 		onNetworkStatus,
 		parentItemId,
