+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/thunks/pasteItems.js
@@ -3,14 +3,14 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import pasteItemAction from '../actions/pasteItem';
+import pasteItemAction from '../actions/pasteItems';
 import {ITEM_ACTIVATION_ORIGINS} from '../config/constants/itemActivationOrigins';
 import FragmentService from '../services/FragmentService';
 import getFirstControlsId from '../utils/getFirstControlsId';
 import sortItemIds from '../utils/sortItemIds';
 import filterSelectedItems from './filterSelectedItems';
 
-export default function pasteItem({
+export default function pasteItems({
 	clipboard = [],
 	parentItemId,
 	selectItems = () => {},
@@ -18,7 +18,7 @@ export default function pasteItem({
 	return (dispatch, getState) => {
 		const {layoutData, segmentsExperienceId} = getState();
 
-		FragmentService.pasteItem({
+		FragmentService.pasteItems({
 			itemIds: sortItemIds(
 				filterSelectedItems(clipboard, layoutData.items),
 				layoutData
