+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/browser/components/page_structure/components/StructureTreeNodeActions.js
@@ -34,7 +34,7 @@ import {
 import {useGetWidgets} from '../../../../../app/contexts/WidgetsContext';
 import deleteItem from '../../../../../app/thunks/deleteItem';
 import duplicateItem from '../../../../../app/thunks/duplicateItem';
-import pasteItem from '../../../../../app/thunks/pasteItem';
+import pasteItems from '../../../../../app/thunks/pasteItems';
 import canBeCopied from '../../../../../app/utils/canBeCopied';
 import canBeDuplicated from '../../../../../app/utils/canBeDuplicated';
 import canBeRemoved from '../../../../../app/utils/canBeRemoved';
@@ -320,7 +320,7 @@ const ActionList = ({item, setActive, setOpenSaveModal}) => {
 						)
 					) {
 						dispatch(
-							pasteItem({
+							pasteItems({
 								clipboard,
 								parentItemId: item.id,
 								selectItems,
